reverted sampleManifest format changes, renamed top level directory from react-marker to html-marker
