+++ src/frontend-samples/html-marker-sample/HtmlMarkerTools.tsx
