+++ src/frontend-samples/html-marker-sample/HtmlMarkerUI.tsx
