+++ src/frontend-samples/html-marker-sample/sampleSpec.ts
