+++ src/frontend-samples/html-marker-sample/utils.tsx
